+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added the ability to sync total_sales when stock value is changed
 * Added the ability to combine product report with its translation
+* Added the ability to combine category report with its translation
 * Fixed database error in sales_be_category reports
 
 ###0.19